RFC: Add proposed Permissions policy for discussion
Please note this is **not** designed to be dictating policy, and may
even be completely wrong.  This is purely supposed to be a starting
point that we can discuss and iterate on during the Summit before
finalizing the policy. We could throw this out tomorrow morning, but I
felt it's better to have a vague starting point that we can discuss than
having to write a whole thing from scratch.