fix: phrasing for attending consecutive meetings. (#450)
* fix: phrasing for attending consecutive meetings.

The motivation for this change that now the Releases WG has meetings in
alternating timezones. Because of this, the prior phrasing "three
consecutive meetings" now has an unindented requirement of attending
meetings in the middle of the night.

* docs: @deermichel suggestion to fix meeting times

Update the meeting time to note that we alternate between meeting times.